back-end/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
pymongo==4.5.0
python-dotenv==1.0.0
pandas==2.2.3
Pillow==11.0.0
pytesseract==0.3.10
requests==2.31.0
openai==1.0.0
instaloader==4.15
waitress==2.1.2
gunicorn==21.2.0
